+++ terrain/voxel_mesh_block.cpp
@@ -42,8 +42,14 @@ void VoxelMeshBlock::set_shadow_casting(RenderingServer::ShadowCastingSetting se
 	}
 }
 
-void VoxelMeshBlock::set_mesh(
-		Ref<Mesh> mesh, GeometryInstance3D::GIMode gi_mode, RenderingServer::ShadowCastingSetting setting) {
+void VoxelMeshBlock::set_render_layers_mask(int mask) {
+	if (_mesh_instance.is_valid()) {
+		_mesh_instance.set_render_layers_mask(mask);
+	}
+}
+
+void VoxelMeshBlock::set_mesh(Ref<Mesh> mesh, GeometryInstance3D::GIMode gi_mode,
+		RenderingServer::ShadowCastingSetting shadow_setting, int render_layers_mask) {
 	// TODO Don't add mesh instance to the world if it's not visible.
 	// I suspect Godot is trying to include invisible mesh instances into the culling process,
 	// which is killing performance when LOD is used (i.e many meshes are in pool but hidden)
@@ -54,7 +60,8 @@ void VoxelMeshBlock::set_mesh(
 			// Create instance if it doesn't exist
 			_mesh_instance.create();
 			_mesh_instance.set_gi_mode(gi_mode);
-			_mesh_instance.set_cast_shadows_setting(setting);
+			_mesh_instance.set_cast_shadows_setting(shadow_setting);
+			_mesh_instance.set_render_layers_mask(render_layers_mask);
 			set_mesh_instance_visible(_mesh_instance, _visible && _parent_visible);
 		}