+++ terrain/voxel_mesh_block.h
@@ -35,8 +35,8 @@ public:
 
 	// Visuals
 
-	void set_mesh(
-			Ref<Mesh> mesh, GeometryInstance3D::GIMode gi_mode, RenderingServer::ShadowCastingSetting shadow_casting);
+	void set_mesh(Ref<Mesh> mesh, GeometryInstance3D::GIMode gi_mode,
+			RenderingServer::ShadowCastingSetting shadow_setting, int render_layers_mask);
 	Ref<Mesh> get_mesh() const;
 	bool has_mesh() const;
 	void drop_mesh();
@@ -48,6 +48,10 @@ public:
 	// Note, ShadowCastingSetting is not stored per block, it is a shared option so we provide it in several functions.
 	// Call this function only if the mesh block already exists and has not changed mesh
 	void set_shadow_casting(RenderingServer::ShadowCastingSetting setting);
+
+	// Note, render layers is not stored per block, it is a shared option so we provide it in several functions.
+	// Call this function only if the mesh block already exists and has not changed mesh
+	void set_render_layers_mask(int mask);
 
 	// Collisions