+++ util/godot/direct_mesh_instance.h
@@ -29,6 +29,7 @@ public:
 	void set_cast_shadows_setting(RenderingServer::ShadowCastingSetting mode);
 	void set_shader_instance_parameter(StringName key, Variant value);
 	void set_gi_mode(GeometryInstance3D::GIMode mode);
+	void set_render_layers_mask(int mask);
 
 	Ref<Mesh> get_mesh() const;